# replace buffer channel pool with sync.Pool