--- a/wire/common.go
+++ b/wire/common.go
@@ -10,6 +10,7 @@ import (
 	"fmt"
 	"io"
 	"math"
+	"sync"
 	"time"
 
 	"github.com/btcsuite/btcd/chaincfg/chainhash"
@@ -45,131 +46,138 @@ var (
 // integers that automatically obtain a buffer from the free list, perform the
 // necessary binary conversion, read from or write to the given io.Reader or
 // io.Writer, and return the buffer to the free list.
-type binaryFreeList chan []byte
+type binaryFreeList struct {
+	sync.Pool
+}
 
 // Borrow returns a byte slice from the free list with a length of 8.  A new
 // buffer is allocated if there are not any available on the free list.
-func (l binaryFreeList) Borrow() []byte {
-	var buf []byte
-	select {
-	case buf = <-l:
-	default:
-		buf = make([]byte, 8)
-	}
-	return buf[:8]
+func (l *binaryFreeList) Borrow() *[]byte {
+	return l.Get().(*[]byte)
 }
 
 // Return puts the provided byte slice back on the free list.  The buffer MUST
 // have been obtained via the Borrow function and therefore have a cap of 8.
-func (l binaryFreeList) Return(buf []byte) {
-	select {
-	case l <- buf:
-	default:
-		// Let it go to the garbage collector.
-	}
+func (l *binaryFreeList) Return(bufp *[]byte) {
+	l.Put(bufp)
 }
 
 // Uint8 reads a single byte from the provided reader using a buffer from the
 // free list and returns it as a uint8.
-func (l binaryFreeList) Uint8(r io.Reader) (uint8, error) {
-	buf := l.Borrow()[:1]
+func (l *binaryFreeList) Uint8(r io.Reader) (uint8, error) {
+	bufp := l.Borrow()
+	buf := (*bufp)[:1]
 	if _, err := io.ReadFull(r, buf); err != nil {
-		l.Return(buf)
+		l.Return(bufp)
 		return 0, err
 	}
 	rv := buf[0]
-	l.Return(buf)
+	l.Return(bufp)
 	return rv, nil
 }
 
 // Uint16 reads two bytes from the provided reader using a buffer from the
 // free list, converts it to a number using the provided byte order, and returns
 // the resulting uint16.
-func (l binaryFreeList) Uint16(r io.Reader, byteOrder binary.ByteOrder) (uint16, error) {
-	buf := l.Borrow()[:2]
+func (l *binaryFreeList) Uint16(r io.Reader, byteOrder binary.ByteOrder) (uint16, error) {
+	bufp := l.Borrow()
+	buf := (*bufp)[:2]
 	if _, err := io.ReadFull(r, buf); err != nil {
-		l.Return(buf)
+		l.Return(bufp)
 		return 0, err
 	}
 	rv := byteOrder.Uint16(buf)
-	l.Return(buf)
+	l.Return(bufp)
 	return rv, nil
 }
 
 // Uint32 reads four bytes from the provided reader using a buffer from the
 // free list, converts it to a number using the provided byte order, and returns
 // the resulting uint32.
-func (l binaryFreeList) Uint32(r io.Reader, byteOrder binary.ByteOrder) (uint32, error) {
-	buf := l.Borrow()[:4]
+func (l *binaryFreeList) Uint32(r io.Reader, byteOrder binary.ByteOrder) (uint32, error) {
+	bufp := l.Borrow()
+	buf := (*bufp)[:4]
 	if _, err := io.ReadFull(r, buf); err != nil {
-		l.Return(buf)
+		l.Return(bufp)
 		return 0, err
 	}
 	rv := byteOrder.Uint32(buf)
-	l.Return(buf)
+	l.Return(bufp)
 	return rv, nil
 }
 
 // Uint64 reads eight bytes from the provided reader using a buffer from the
 // free list, converts it to a number using the provided byte order, and returns
 // the resulting uint64.
-func (l binaryFreeList) Uint64(r io.Reader, byteOrder binary.ByteOrder) (uint64, error) {
-	buf := l.Borrow()[:8]
+func (l *binaryFreeList) Uint64(r io.Reader, byteOrder binary.ByteOrder) (uint64, error) {
+	bufp := l.Borrow()
+	buf := (*bufp)[:8]
 	if _, err := io.ReadFull(r, buf); err != nil {
-		l.Return(buf)
+		l.Return(bufp)
 		return 0, err
 	}
 	rv := byteOrder.Uint64(buf)
-	l.Return(buf)
+	l.Return(bufp)
 	return rv, nil
 }
 
 // PutUint8 copies the provided uint8 into a buffer from the free list and
 // writes the resulting byte to the given writer.
-func (l binaryFreeList) PutUint8(w io.Writer, val uint8) error {
-	buf := l.Borrow()[:1]
+func (l *binaryFreeList) PutUint8(w io.Writer, val uint8) error {
+	bufp := l.Borrow()
+	buf := (*bufp)[:1]
 	buf[0] = val
 	_, err := w.Write(buf)
-	l.Return(buf)
+	l.Return(bufp)
 	return err
 }
 
 // PutUint16 serializes the provided uint16 using the given byte order into a
 // buffer from the free list and writes the resulting two bytes to the given
 // writer.
-func (l binaryFreeList) PutUint16(w io.Writer, byteOrder binary.ByteOrder, val uint16) error {
-	buf := l.Borrow()[:2]
+func (l *binaryFreeList) PutUint16(w io.Writer, byteOrder binary.ByteOrder, val uint16) error {
+	bufp := l.Borrow()
+	buf := (*bufp)[:2]
 	byteOrder.PutUint16(buf, val)
 	_, err := w.Write(buf)
-	l.Return(buf)
+	l.Return(bufp)
 	return err
 }
 
 // PutUint32 serializes the provided uint32 using the given byte order into a
 // buffer from the free list and writes the resulting four bytes to the given
 // writer.
-func (l binaryFreeList) PutUint32(w io.Writer, byteOrder binary.ByteOrder, val uint32) error {
-	buf := l.Borrow()[:4]
+func (l *binaryFreeList) PutUint32(w io.Writer, byteOrder binary.ByteOrder, val uint32) error {
+	bufp := l.Borrow()
+	buf := (*bufp)[:4]
 	byteOrder.PutUint32(buf, val)
 	_, err := w.Write(buf)
-	l.Return(buf)
+	l.Return(bufp)
 	return err
 }
 
 // PutUint64 serializes the provided uint64 using the given byte order into a
 // buffer from the free list and writes the resulting eight bytes to the given
 // writer.
-func (l binaryFreeList) PutUint64(w io.Writer, byteOrder binary.ByteOrder, val uint64) error {
-	buf := l.Borrow()[:8]
+func (l *binaryFreeList) PutUint64(w io.Writer, byteOrder binary.ByteOrder, val uint64) error {
+	bufp := l.Borrow()
+	buf := (*bufp)[:8]
 	byteOrder.PutUint64(buf, val)
 	_, err := w.Write(buf)
-	l.Return(buf)
+	l.Return(bufp)
 	return err
 }
 
 // binarySerializer provides a free list of buffers to use for serializing and
 // deserializing primitive integer values to and from io.Readers and io.Writers.
-var binarySerializer binaryFreeList = make(chan []byte, binaryFreeListMaxItems)
+var binarySerializer = &binaryFreeList{
+	Pool: sync.Pool{
+		New: func() interface{} {
+			b := make([]byte, 8)
+			return &b
+		},
+	},
+}
 
 // errNonCanonicalVarInt is the common format string used for non-canonically
 // encoded variable length integer errors.