@davecgh 2017-10-12
Thanks for the PR, however, as you noted, the `sync.Pool` entries get removed at any time which defeats the purpose of why the code is maintain its own free list to begin with. See https://github.com/btcsuite/btcd/pull/684 for more details.